feat: [TG-7] 메서드 이동 및 일정 조회 api 기능
