+++ frontend/src/mapper/managementTargetMapper.js
@@ -1,3 +1,5 @@
+import { jsonToDatetime } from "./simpleMapper";
+
 // react object name : json name
 export function jsonToTargetInfomation(data) {
     return {
@@ -12,23 +14,6 @@ export function jsonToTargetInfomation(data) {
         "createdAt": jsonToDatetime(data.createdAt),
         "updatedAt": jsonToDatetime(data.updatedAt),
     };
-}
-
-export function jsonToHistoryOfSchedule(data) {
-    return data.map(d => mapToScheduleObject(d));
-}
-
-function mapToScheduleObject(data) {
-    return {
-        "scheduleId": data.schedulId,
-        "title": data.title,
-        "content": data.content,
-        "startDate": jsonToDatetime(data.startDate),
-        "endDate": jsonToDatetime(data.endDate),
-        "createdAt": jsonToDatetime(data.createdAt),
-        "updatedAt": jsonToDatetime(data.updatedAt),
-        "activated": data.activated,
-    }
 }
 
 export function jsonToUsageForm(data) {
@@ -56,8 +41,4 @@ export function jsonToUsageForm(data) {
             }
         ]
     };
-}
-
-function jsonToDatetime(str) {
-    return new Date(str);
 }
